backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-multipart==0.0.9
pydantic==2.7.4
pdfminer.six==20231228
PyMuPDF==1.24.9
numpy>=1.26,<2.0
fastembed==0.3.4
google-generativeai==0.7.2
azure-cognitiveservices-speech>=1.40.0
requests==2.31.0

